Update to the README.md to include section numbers
I found this made it much easier to understand the structure of the document as I was reading it. It makes it clear which sections are subsections, and sub-subsections.

I also changed some of the header levels under the assumption that the table of contents was correct about the organization scheme (e.g. 3. Revoking account access was originally a h3 in the document, despite the table of contents specifying it as an h2)